requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Migrate
reportlab
pikepdf
structlog
PyJWT